piece had more than one  responsibility , so single responsibility principle was violated, Piece movement strategy was outed, created new MovementStrategy , and coressponding movoments, also GetDiagonals and getLinears was outed in utility class
