+++ src/main/java/chess/model/Bishop.java
@@ -1,6 +1,8 @@
 package chess.model;
 
 import chess.model.enums.PieceColor;
+import chess.movement.MovementStrategy;
+import chess.movement.StandardBishopMovement;
 
 import java.util.List;
 
@@ -9,13 +11,10 @@ public class Bishop extends Piece {
     public Bishop(PieceColor color, Square initSq, String img_file) {
         super(color, initSq, img_file);
     }
-    
+
     @Override
-    public List<Square> getLegalMoves(Board b) {
-        Square[][] board = b.getSquareArray();
-        int x = this.getPosition().getXNum();
-        int y = this.getPosition().getYNum();
-        
-        return getDiagonalOccupations(board, x, y);
+    protected MovementStrategy GetMovementStrategy() {
+        return new StandardBishopMovement(this);
     }
+
 }
