+++ src/main/java/chess/model/King.java
@@ -1,6 +1,8 @@
 package chess.model;
 
 import chess.model.enums.PieceColor;
+import chess.movement.MovementStrategy;
+import chess.movement.StandardKingMovement;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -12,31 +14,9 @@ public class King extends Piece {
     }
 
     @Override
-    public List<Square> getLegalMoves(Board b) {
-LinkedList<Square> legalMoves = new LinkedList<Square>();
-        
-        Square[][] board = b.getSquareArray();
-        
-        int x = this.getPosition().getXNum();
-        int y = this.getPosition().getYNum();
-        
-        for (int i = 1; i > -2; i--) {
-            for (int k = 1; k > -2; k--) {
-                if(!(i == 0 && k == 0)) {
-                    try {
-                        if(!board[y + k][x + i].isOccupied() || 
-                                board[y + k][x + i].getOccupyingPiece().getColor() 
-                                != this.getColor()) {
-                            legalMoves.add(board[y + k][x + i]);
-                        }
-                    } catch (ArrayIndexOutOfBoundsException e) {
-                        continue;
-                    }
-                }
-            }
-        }
-        
-        return legalMoves;
+    protected MovementStrategy GetMovementStrategy() {
+        return new StandardKingMovement(this);
     }
+
 
 }
