+++ src/main/java/chess/model/Knight.java
@@ -1,6 +1,8 @@
 package chess.model;
 
 import chess.model.enums.PieceColor;
+import chess.movement.MovementStrategy;
+import chess.movement.StandardKnightMovement;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -12,28 +14,9 @@ public class Knight extends Piece {
     }
 
     @Override
-    public List<Square> getLegalMoves(Board b) {
-        LinkedList<Square> legalMoves = new LinkedList<Square>();
-        Square[][] board = b.getSquareArray();
-        
-        int x = this.getPosition().getXNum();
-        int y = this.getPosition().getYNum();
-        
-        for (int i = 2; i > -3; i--) {
-            for (int k = 2; k > -3; k--) {
-                if(Math.abs(i) == 2 ^ Math.abs(k) == 2) {
-                    if (k != 0 && i != 0) {
-                        try {
-                            legalMoves.add(board[y + k][x + i]);
-                        } catch (ArrayIndexOutOfBoundsException e) {
-                            continue;
-                        }
-                    }
-                }
-            }
-        }
-        
-        return legalMoves;
+    protected MovementStrategy GetMovementStrategy() {
+        return new StandardKnightMovement(this);
     }
+
 
 }
