+++ src/main/java/chess/model/Pawn.java
@@ -1,86 +1,27 @@
 package chess.model;
 
 import chess.model.enums.PieceColor;
+import chess.movement.MovementStrategy;
+import chess.movement.StandardPawnMovement;
 
 import java.util.List;
 import java.util.LinkedList;
 
 public class Pawn extends Piece {
-    private boolean wasMoved;
-    
+
+
     public Pawn(PieceColor color, Square initSq, String img_file) {
         super(color, initSq, img_file);
     }
-    
+
     @Override
-    public boolean move(Square fin) {
-        boolean b = super.move(fin);
-        wasMoved = true;
-        return b;
+    protected MovementStrategy GetMovementStrategy() {
+        return new StandardPawnMovement(this,false);
     }
 
     @Override
-    public List<Square> getLegalMoves(Board b) {
-        LinkedList<Square> legalMoves = new LinkedList<Square>();
-        
-        Square[][] board = b.getSquareArray();
-        
-        int x = this.getPosition().getXNum();
-        int y = this.getPosition().getYNum();
-        PieceColor c = this.getColor();
-        
-        if (c == PieceColor.BLACK) {
-            if (!wasMoved) {
-                if (!board[y+2][x].isOccupied()) {
-                    legalMoves.add(board[y+2][x]);
-                }
-            }
-            
-            if (y+1 < 8) {
-                if (!board[y+1][x].isOccupied()) {
-                    legalMoves.add(board[y+1][x]);
-                }
-            }
-            
-            if (x+1 < 8 && y+1 < 8) {
-                if (board[y+1][x+1].isOccupied()) {
-                    legalMoves.add(board[y+1][x+1]);
-                }
-            }
-                
-            if (x-1 >= 0 && y+1 < 8) {
-                if (board[y+1][x-1].isOccupied()) {
-                    legalMoves.add(board[y+1][x-1]);
-                }
-            }
-        }
-        
-        if (c == PieceColor.WHITE) {
-            if (!wasMoved) {
-                if (!board[y-2][x].isOccupied()) {
-                    legalMoves.add(board[y-2][x]);
-                }
-            }
-            
-            if (y-1 >= 0) {
-                if (!board[y-1][x].isOccupied()) {
-                    legalMoves.add(board[y-1][x]);
-                }
-            }
-            
-            if (x+1 < 8 && y-1 >= 0) {
-                if (board[y-1][x+1].isOccupied()) {
-                    legalMoves.add(board[y-1][x+1]);
-                }
-            }
-                
-            if (x-1 >= 0 && y-1 >= 0) {
-                if (board[y-1][x-1].isOccupied()) {
-                    legalMoves.add(board[y-1][x-1]);
-                }
-            }
-        }
-        
-        return legalMoves;
+    public boolean move(Square fin) {
+        setMovementStrategy(new StandardPawnMovement(this,true));
+        return super.move(fin);
     }
 }
