+++ src/main/java/chess/model/Piece.java
@@ -1,6 +1,10 @@
 package chess.model;
 
 import chess.model.enums.PieceColor;
+import chess.movement.MovementStrategy;
+import chess.util.PieceImageLoader;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
@@ -14,175 +18,55 @@ public abstract class Piece {
     private final PieceColor color;
     private Square currentSquare;
     private BufferedImage img;
-    
+    @Setter
+    private MovementStrategy movementStrategy;
+
     public Piece(PieceColor color, Square initSq, String img_file) {
         this.color = color;
         this.currentSquare = initSq;
-        
-        try {
-            if (this.img == null) {
-              this.img = ImageIO.read(getClass().getResource(img_file));
-            }
-          } catch (IOException e) {
-            System.out.println("File not found: " + e.getMessage());
-          }
+
+        this.img = PieceImageLoader.loadImage(img_file);
+        this.movementStrategy = GetMovementStrategy();
     }
-    
+
     public boolean move(Square fin) {
         Piece occup = fin.getOccupyingPiece();
-        
+
         if (occup != null) {
             if (occup.getColor() == this.color) return false;
             else fin.capture(this);
         }
-        
+
         currentSquare.removePiece();
         this.currentSquare = fin;
         currentSquare.put(this);
         return true;
     }
-    
+
+    protected abstract MovementStrategy GetMovementStrategy();
+
+
+
     public Square getPosition() {
         return currentSquare;
     }
-    
+
     public void setPosition(Square sq) {
         this.currentSquare = sq;
     }
-    
+
     public PieceColor getColor() {
         return color;
     }
-    
+
     public Image getImage() {
         return img;
     }
-    
-//    public void draw(Graphics g) {
-//        int x = currentSquare.getX();
-//        int y = currentSquare.getY();
-//
-//        g.drawImage(this.img, x, y, null);
-//    }
-    
-    public int[] getLinearOccupations(Square[][] board, int x, int y) {
-        int lastYabove = 0;
-        int lastXright = 7;
-        int lastYbelow = 7;
-        int lastXleft = 0;
-        
-        for (int i = 0; i < y; i++) {
-            if (board[i][x].isOccupied()) {
-                if (board[i][x].getOccupyingPiece().getColor() != this.color) {
-                    lastYabove = i;
-                } else lastYabove = i + 1;
-            }
-        }
 
-        for (int i = 7; i > y; i--) {
-            if (board[i][x].isOccupied()) {
-                if (board[i][x].getOccupyingPiece().getColor() != this.color) {
-                    lastYbelow = i;
-                } else lastYbelow = i - 1;
-            }
-        }
-
-        for (int i = 0; i < x; i++) {
-            if (board[y][i].isOccupied()) {
-                if (board[y][i].getOccupyingPiece().getColor() != this.color) {
-                    lastXleft = i;
-                } else lastXleft = i + 1;
-            }
-        }
-
-        for (int i = 7; i > x; i--) {
-            if (board[y][i].isOccupied()) {
-                if (board[y][i].getOccupyingPiece().getColor() != this.color) {
-                    lastXright = i;
-                } else lastXright = i - 1;
-            }
-        }
-        
-        int[] occups = {lastYabove, lastYbelow, lastXleft, lastXright};
-        
-        return occups;
+    public List<Square> getLegalMoves(Board b) {
+        return movementStrategy.getLegalMoves(b);
     }
-    
-    public List<Square> getDiagonalOccupations(Square[][] board, int x, int y) {
-        LinkedList<Square> diagOccup = new LinkedList<Square>();
-        
-        int xNW = x - 1;
-        int xSW = x - 1;
-        int xNE = x + 1;
-        int xSE = x + 1;
-        int yNW = y - 1;
-        int ySW = y + 1;
-        int yNE = y - 1;
-        int ySE = y + 1;
-        
-        while (xNW >= 0 && yNW >= 0) {
-            if (board[yNW][xNW].isOccupied()) {
-                if (board[yNW][xNW].getOccupyingPiece().getColor() == this.color) {
-                    break;
-                } else {
-                    diagOccup.add(board[yNW][xNW]);
-                    break;
-                }
-            } else {
-                diagOccup.add(board[yNW][xNW]);
-                yNW--;
-                xNW--;
-            }
-        }
-        
-        while (xSW >= 0 && ySW < 8) {
-            if (board[ySW][xSW].isOccupied()) {
-                if (board[ySW][xSW].getOccupyingPiece().getColor() == this.color) {
-                    break;
-                } else {
-                    diagOccup.add(board[ySW][xSW]);
-                    break;
-                }
-            } else {
-                diagOccup.add(board[ySW][xSW]);
-                ySW++;
-                xSW--;
-            }
-        }
-        
-        while (xSE < 8 && ySE < 8) {
-            if (board[ySE][xSE].isOccupied()) {
-                if (board[ySE][xSE].getOccupyingPiece().getColor() == this.color) {
-                    break;
-                } else {
-                    diagOccup.add(board[ySE][xSE]);
-                    break;
-                }
-            } else {
-                diagOccup.add(board[ySE][xSE]);
-                ySE++;
-                xSE++;
-            }
-        }
-        
-        while (xNE < 8 && yNE >= 0) {
-            if (board[yNE][xNE].isOccupied()) {
-                if (board[yNE][xNE].getOccupyingPiece().getColor() == this.color) {
-                    break;
-                } else {
-                    diagOccup.add(board[yNE][xNE]);
-                    break;
-                }
-            } else {
-                diagOccup.add(board[yNE][xNE]);
-                yNE--;
-                xNE++;
-            }
-        }
-        
-        return diagOccup;
-    }
-    
-    // No implementation, to be implemented by each subclass
-    public abstract List<Square> getLegalMoves(Board b);
+
+
+
 }
