+++ src/main/java/chess/model/Rook.java
@@ -1,6 +1,8 @@
 package chess.model;
 
 import chess.model.enums.PieceColor;
+import chess.movement.MovementStrategy;
+import chess.movement.StandardRookMovement;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -12,24 +14,9 @@ public class Rook extends Piece {
     }
 
     @Override
-    public List<Square> getLegalMoves(Board b) {
-        LinkedList<Square> legalMoves = new LinkedList<Square>();
-        Square[][] board = b.getSquareArray();
-        
-        int x = this.getPosition().getXNum();
-        int y = this.getPosition().getYNum();
-        
-        int[] occups = getLinearOccupations(board, x, y);
-        
-        for (int i = occups[0]; i <= occups[1]; i++) {
-            if (i != y) legalMoves.add(board[i][x]);
-        }
-        
-        for (int i = occups[2]; i <= occups[3]; i++) {
-            if (i != x) legalMoves.add(board[y][i]);
-        }
-        
-        return legalMoves;
+    protected MovementStrategy GetMovementStrategy() {
+        return new StandardRookMovement(this);
     }
+
 
 }
