--- src/main/java/chess/movement/StandartBishopMovement.java
@@ -1,4 +1,0 @@
-package chess.movement;
-
-public class StandartBishopMovement {
-}
